rename app to student system
