+++ app/src/androidTest/java/com/example/studentsystem/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.testapp;
+package com.example.studentsystem;
 
 import android.content.Context;
 
@@ -22,6 +22,6 @@ public class ExampleInstrumentedTest {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
 
-        assertEquals("com.example.testapp", appContext.getPackageName());
+        assertEquals("com.example.studentsystem", appContext.getPackageName());
     }
 }
