+++ app/src/main/java/com/example/studentsystem/RegisterFragment.java
@@ -1,4 +1,4 @@
-package com.example.testapp;
+package com.example.studentsystem;
 
 
 import android.os.Bundle;
