+++ app/src/main/java/com/example/studentsystem/ui/gallery/GalleryFragment.java
@@ -1,4 +1,4 @@
-package com.example.testapp.ui.gallery;
+package com.example.studentsystem.ui.gallery;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -12,7 +12,7 @@ import androidx.fragment.app.Fragment;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
 
-import com.example.testapp.R;
+import com.example.studentsystem.R;
 
 public class GalleryFragment extends Fragment {
