+++ app/src/main/java/com/example/studentsystem/ui/share/ShareViewModel.java
@@ -1,16 +1,16 @@
-package com.example.testapp.ui.home;
+package com.example.studentsystem.ui.share;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class HomeViewModel extends ViewModel {
+public class ShareViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public HomeViewModel() {
+    public ShareViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
+        mText.setValue("This is share fragment");
     }
 
     public LiveData<String> getText() {
