+++ app/src/main/java/com/example/studentsystem/ui/slideshow/SlideshowViewModel.java
@@ -1,16 +1,16 @@
-package com.example.testapp.ui.tools;
+package com.example.studentsystem.ui.slideshow;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class ToolsViewModel extends ViewModel {
+public class SlideshowViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public ToolsViewModel() {
+    public SlideshowViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is tools fragment");
+        mText.setValue("This is slideshow fragment");
     }
 
     public LiveData<String> getText() {
